include macros
Added a new macro type for includes. These macros can be used anywhere
in the preprocessed code. They are invoked with the 'minclude' keyword
and must have the include <name>...end include syntax
This commit also brings an update to the way macros are created and stored
by adding subclasses.