Updated makefile and binary to use .f90 file endings.
ifort cannot handle .f03 out of the box and files with this ending need
to be passed in a very cumbersome way. .f90 is easier handled by all compilers.
we should be able to pass flags to other compilers to indicate that we're
dealing with Fortran 2003 code (maybe provided by ppm/settings).