Replaced recursive macro calling hacks by recursive preproc
The original implementation used extra regexp and erb code to allow for
recursive macro calling. This could not be easily expanded to allow arbitrary
macro recursions (foreach macros in function macros, etc.).
This code is now replaced by a recursive call to the preprocessor. This
conflicts with the singleton preprocessor, therefor the preproc was enhanced
with a state stack to allow for preproc hierarchies